Unit tests sprint 1

+++ src/main/java/controller/AllController.java
@@ -2,10 +2,13 @@ package controller;
 
 import model.*;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import service.AllService;
-import service.StudentService;
+import org.springframework.web.util.UriComponentsBuilder;
 
+import service.AllService;
 import java.util.List;
 
 /**
@@ -18,21 +21,6 @@ public class AllController {
     @Autowired
     private AllService allService;
 
-    @RequestMapping(path = "products", method = RequestMethod.POST)
-    public Product addProduct(@RequestBody Product product){
-        return allService.saveProduct(product);
-    }
-
-    @RequestMapping(path = "receipts", method = RequestMethod.POST)
-    public Receipt addReceipt(@RequestBody Receipt receipt){
-        return allService.saveReceipt(receipt);
-    }
-
-    @RequestMapping(path = "receiptDetails", method = RequestMethod.POST)
-    public ReceiptDetail addReceipt(@RequestBody ReceiptDetail receiptDetail){
-        return allService.saveReceiptDetail(receiptDetail);
-    }
-
     //Assign Controllers
 
     //Business
@@ -42,23 +30,50 @@ public class AllController {
     }
 
     @RequestMapping(path = "businesses/{businessId}", method = RequestMethod.GET)
-    public Business getBusiness(@PathVariable int businessId){
-        return allService.getBusiness(businessId);
+    public ResponseEntity<Business> getBusiness(@PathVariable("businessId") int businessId){
+    	Business business = allService.getBusiness(businessId);
+
+        if (business == null){
+            return new ResponseEntity<Business>(HttpStatus.NOT_FOUND);
+        }
+
+        return new ResponseEntity<Business>(business, HttpStatus.OK);
     }
 
     @RequestMapping(path = "businesses", method = RequestMethod.POST)
-    public Business addBusiness(@RequestBody Business business){
-        return allService.saveBusiness(business);
+    public ResponseEntity<Void> saveBusiness(@RequestBody Business business, UriComponentsBuilder ucBuilder){
+
+        allService.saveBusiness(business);
+
+        HttpHeaders headers = new HttpHeaders();
+        headers.setLocation(ucBuilder.path("/businesses/{id}").buildAndExpand(business.getId()).toUri());
+        return new ResponseEntity<Void>(headers, HttpStatus.CREATED);
+        
     }
 
     @RequestMapping(path = "businesses", method = RequestMethod.PUT)
-    public void updateBusiness(@RequestBody  Business business){
+    public ResponseEntity<Void> updateBusiness(@RequestBody Business business){
+        Business businessToBeUpdated = allService.getBusiness(business.getId());
+
+        if (businessToBeUpdated == null){
+            return new ResponseEntity<Void>(HttpStatus.NOT_FOUND);
+        }
+        
         allService.updateBusiness(business);
+
+        return new ResponseEntity<Void>(HttpStatus.OK);
     }
 
     @RequestMapping(path = "businesses/{businessId}", method = RequestMethod.DELETE)
-    public void deleteBusiness(@PathVariable int businessId){
+    public ResponseEntity<Void> deleteBusiness(@PathVariable("businessId") int businessId){
+    	Business business = allService.getBusiness(businessId);
+
+        if (business == null){
+            return new ResponseEntity<Void>(HttpStatus.NOT_FOUND);
+        }
+
         allService.deleteBusiness(businessId);
+        return new ResponseEntity<Void>(HttpStatus.OK);
     }
 
     //Customer
@@ -68,22 +83,53 @@ public class AllController {
     }
 
     @RequestMapping(path = "customers/{customerId}", method = RequestMethod.GET)
-    public Customer getCustomer(@PathVariable int customerId){
-        return allService.getCustomer(customerId);
+    public ResponseEntity<Customer> getCustomer(@PathVariable("customerId") int customerId){
+    	Customer customer = allService.getCustomer(customerId);
+
+        if (customer == null){
+            return new ResponseEntity<Customer>(HttpStatus.NOT_FOUND);
+        }
+
+        return new ResponseEntity<Customer>(customer, HttpStatus.OK);
     }
 
     @RequestMapping(path = "customers", method = RequestMethod.POST)
-    public Customer addCustomer(@RequestBody Customer customer){
-        return allService.saveCustomer(customer);
+    public ResponseEntity<Void> saveCustomer(@RequestBody Customer customer, UriComponentsBuilder ucBuilder){
+
+        if (allService.customerExists(customer.getUsername())){
+            return new ResponseEntity<Void>(HttpStatus.CONFLICT);
+        }
+
+        allService.saveCustomer(customer);
+
+        HttpHeaders headers = new HttpHeaders();
+        headers.setLocation(ucBuilder.path("/customers/{id}").buildAndExpand(customer.getId()).toUri());
+        return new ResponseEntity<Void>(headers, HttpStatus.CREATED);
+        
     }
 
     @RequestMapping(path = "customers", method = RequestMethod.PUT)
-    public void updateCustomer(@RequestBody  Customer customer){
+    public ResponseEntity<Void> updateCustomer(@RequestBody Customer customer){
+        Customer customerToBeUpdated = allService.getCustomer(customer.getId());
+
+        if (customerToBeUpdated == null){
+            return new ResponseEntity<Void>(HttpStatus.NOT_FOUND);
+        }
+        
         allService.updateCustomer(customer);
+
+        return new ResponseEntity<Void>(HttpStatus.OK);
     }
 
     @RequestMapping(path = "customers/{customerId}", method = RequestMethod.DELETE)
-    public void deleteCustomer(@PathVariable int customerId){
+    public ResponseEntity<Void> deleteCustomer(@PathVariable("customerId") int customerId){
+    	Customer customer = allService.getCustomer(customerId);
+
+        if (customer == null){
+            return new ResponseEntity<Void>(HttpStatus.NOT_FOUND);
+        }
+
         allService.deleteCustomer(customerId);
+        return new ResponseEntity<Void>(HttpStatus.OK);
     }
 }
